backend/requirements.txt
# AI Coordination Agent - Backend Dependencies
# Core Framework - Flask (Updated from FastAPI)
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1

# Database & ORM
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9           # PostgreSQL driver
pgvector==0.2.4                  # Vector database support

# Configuration & Environment
python-dotenv==1.0.0

# HTTP & API
httpx==0.25.2
requests==2.31.0

# Authentication & Security
pyjwt==2.8.0                     # JWT token handling
bcrypt==4.1.2                    # Password hashing
python-multipart==0.0.6

# AI & Machine Learning
openai==1.3.7
numpy>=1.26.0

# Database Drivers (synchronous fallback)
mysql-connector-python==8.2.0
psycopg2-binary==2.9.9

# Data Validation & Serialization
marshmallow==3.20.1

# Vector Database - RAG Pipeline
chromadb==0.4.18
sentence-transformers==2.2.2

# Communication Services (Optional)
python-telegram-bot==20.7         # Telegram
yagmail==0.15.293                 # Email

# Development & Testing
pytest==7.4.0
pytest-asyncio==0.21.1
httpx==0.25.2                     # For testing FastAPI
black==23.9.0
flake8==6.1.0

# Production
redis==5.0.0                      # Caching
celery==5.3.0                     # Background tasks